examples/simple-case/pyproject.toml
[project]
name = "simple-case"
version = "0.1.0"
description = "Example demonstrating simple span-to-transcript cases"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.6",
    "inspect-ai",
    "inspect-scout",
    "langchain",
    "langchain-openai",
    "langgraph",
    "scout-otel",
    "click",
    "python-dotenv",
]

[tool.uv.sources]
scout-otel = { workspace = true }

[project.scripts]
simple-case = "simple_case.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/simple_case"]
